refactor: Move block_timer.h to silo/common directory
